renamed parsing into paring_gapc
